+++ vogs/src/push_to_stack.c
@@ -12,7 +12,7 @@
 
 #include "push_swap.h"
 
-t_block		*push_to_stack(int number, t_block *stack)
+t_block		*push_to_stack(long long number, t_block *stack)
 {
 	t_block	*temp;
